TotalSize 구하는 SQL 위치 변경

+++ src/main/java/org/openflamingo/hadoop/repository/AprioriRepository.java
@@ -9,5 +9,5 @@ package org.openflamingo.hadoop.repository;
 public interface AprioriRepository {
     public void saveCadidate(final String key, final String value) throws Exception;
     public void saveSupport(String key, int support) throws Exception;
-    public void saveTotalSize(int size) throws Exception;
+    public void saveTotalSize(long size) throws Exception;
 }
